% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{as_bench_mark}
\alias{as_bench_mark}
\title{Coerce to a bench mark object Bench mark objects}
\usage{
as_bench_mark(x)
}
\arguments{
\item{x}{Object to be coerced}
}
\description{
This is typically needed only if you are performing additional manipulations
after calling \code{\link[bench:mark]{bench::mark()}}.
}
