% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-package.R
\docType{package}
\name{bench-package}
\alias{bench}
\alias{bench-package}
\title{bench: High Precision Timing of R Expressions}
\description{
Tools to accurately benchmark and analyze execution times for R expressions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimhester/bench}
  \item Report bugs at \url{https://github.com/jimhester/bench/issues}
}

}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com}

}
\keyword{internal}
