% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytes.R
\name{bench_bytes_trans}
\alias{bench_bytes_trans}
\title{Benchmark time transformation}
\usage{
bench_bytes_trans(base = 2)
}
\arguments{
\item{base}{base of logarithm}
}
\description{
This both log transforms the times and formats the labels as a \code{bench_time}
object.
}
\keyword{internal}
