% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_time.R
\name{bench_max_memory}
\alias{bench_max_memory}
\title{Returns the maximum amount of memory used by the current process}
\usage{
bench_max_memory()
}
\description{
On POSIX systems this corresponds to
\href{http://man7.org/linux/man-pages/man2/getrusage.2.html}{rusage.ru_maxrss}
and on Windows
\href{https://docs.microsoft.com/en-us/windows/win32/api/psapi/ns-psapi-process_memory_counters}{PROCESS_MEMORY_COUNTERS.PeakWorkingSetSize}
}
