% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_time.R
\name{bench_memory}
\alias{bench_memory}
\title{Measure memory that an expression used.}
\usage{
bench_memory(expr)
}
\arguments{
\item{expr}{A expression to be measured.}
}
\value{
A tibble with two columns
\itemize{
\item The total amount of memory allocated
\item The raw memory allocations as parsed by \code{\link[profmem:readRprofmem]{profmem::readRprofmem()}}
}
}
\description{
Measure memory that an expression used.
}
\examples{
bench_memory(1 + 1:10000)
}
