% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_time.R, R/time.R
\name{bench_time}
\alias{bench_time}
\alias{system_time}
\alias{as_bench_time}
\title{Measure Process CPU and real time that an expression used.}
\usage{
bench_time(expr)

as_bench_time(x)
}
\arguments{
\item{expr}{A expression to be timed.}

\item{x}{A numeric or character vector. Character representations can use
shorthand sizes (see examples).}
}
\value{
A \code{bench_time} object with two values.
\itemize{
\item \code{process} - The process CPU usage of the expression evaluation.
\item \code{real} - The wallclock time of the expression evaluation.
}
}
\description{
Construct, manipulate and display vectors of elapsed times in seconds. These
are numeric vectors, so you can compare them numerically, but they can also
be compared to human readable values such as '10ms'.
}
\details{
On some systems (such as macOS) the process clock has lower
precision than the realtime clock, as a result there may be cases where the
process time is larger than the real time for fast expressions.
}
\examples{
# This will use ~.5 seconds of real time, but very little process time.
bench_time(Sys.sleep(.5))
bench_time("1")
bench_time("1ns")
bench_time("1us")
bench_time("1ms")
bench_time("1s")

bench_time("100ns") < "1ms"

sum(bench_time(c("1MB", "5MB", "500KB")))
}
\seealso{
\code{\link[=bench_memory]{bench_memory()}} To measure memory allocations for a given expression.
}
