% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{bench_time_trans}
\alias{bench_time_trans}
\title{Benchmark time transformation}
\usage{
bench_time_trans(base = 10)
}
\arguments{
\item{base}{base of logarithm}
}
\description{
This both log transforms the times and formats the labels as a \code{bench_time}
object.
}
\keyword{internal}
