% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_benchmarking.R
\name{cb_fetch}
\alias{cb_fetch}
\title{Push or fetch continuous benchmark notes from a git remote}
\usage{
cb_fetch(remote = "origin")
}
\arguments{
\item{remote}{The git remote to use, defaults to 'origin'.}
}
\description{
By default the git client does not push or fetch notes from remotes,
\code{\link[=cb_fetch]{cb_fetch()}} and \code{\link[=cb_push]{cb_push()}} can be used to do this.
}
\seealso{
Other cb: 
\code{\link{cb_run}()}
}
\concept{cb}
