% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_benchmarking.R
\name{cb_plot_time}
\alias{cb_plot_time}
\title{Plot the execution time for continuous benchmarks}
\usage{
cb_plot_time(x, n = 25)
}
\arguments{
\item{x}{Continuous benchmark data, as obtained from \code{\link[=cb_read]{cb_read()}}.}

\item{n}{The number of previous commits with benchmarks you want to display.}
}
\description{
Plot the execution time for continuous benchmarks
}
