% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_benchmarking.R
\name{cb_read}
\alias{cb_read}
\title{Read continuous benchmark data from the git log}
\usage{
cb_read(additional_columns = NULL)
}
\description{
Note if the benchmarks were run on a remote system you may need to fetch the
data locally first with \code{cb_fetch()}.
}
\examples{
\dontrun{
cb_read(additional_columns=c("tree_hash" = "\%T", "author_email" = "\%ae"))
}
}
\seealso{
Other cb: 
\code{\link{cb_fetch}()},
\code{\link{cb_run}()}
}
\concept{cb}
