% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_benchmarking.R
\name{fetch_benchmark_notes}
\alias{fetch_benchmark_notes}
\alias{push_benchmark_notes}
\title{Push or fetch benchmark notes from a git remote}
\usage{
fetch_benchmark_notes(remote = "origin")

push_benchmark_notes(remote = "origin")
}
\arguments{
\item{remote}{The git remote to use, defaults to 'origin'.}
}
\description{
By default the git client does not push or fetch notes from remotes,
\code{\link[=fetch_benchmark_notes]{fetch_benchmark_notes()}} and \code{\link[=push_benchmark_notes]{push_benchmark_notes()}} can be used to do
this explicitly.
}
