% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{scale_bench_expr}
\alias{scale_bench_expr}
\alias{scale_x_bench_expr}
\alias{scale_y_bench_expr}
\alias{scale_colour_bench_expr}
\alias{scale_color_bench_expr}
\title{Position and color scales for bench_expr data}
\usage{
scale_x_bench_expr(...)

scale_y_bench_expr(...)

scale_colour_bench_expr(
  palette = scales::hue_pal(...),
  ...,
  aesthetics = "colour"
)

scale_color_bench_expr(
  palette = scales::hue_pal(...),
  ...,
  aesthetics = "colour"
)
}
\description{
Default scales for the \code{bench_expr} class, these are added to plots using
\code{bench_expr} objects automatically.
}
\keyword{internal}
