% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytes.R, R/time.R
\name{scale_bench_time}
\alias{scale_bench_time}
\alias{scale_x_bench_bytes}
\alias{scale_y_bench_bytes}
\alias{scale_x_bench_time}
\alias{scale_y_bench_time}
\title{Position scales for bench_time data}
\usage{
scale_x_bench_bytes(base = 10, ...)

scale_y_bench_bytes(base = 10, ...)

scale_x_bench_time(base = 10, ...)

scale_y_bench_time(base = 10, ...)
}
\arguments{
\item{base}{The base of the logarithm, if \code{NULL} instead use a
non-logarithmic scale.}
}
\description{
Default scales for the \code{bench_time} class, these are added to plots using
\code{bench_time} objects automatically.

Default scales for the \code{bench_time} class, these are added to plots using
\code{bench_time} objects automatically.
}
\keyword{internal}
