% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_benchmarking.R
\name{cb_read}
\alias{cb_read}
\title{Read continuous benchmark data from the git log}
\usage{
cb_read(path = ".", additional_columns = NULL)
}
\arguments{
\item{path}{A path to a package or within a package.}

\item{additional_columns}{A named list of additional columns to include. The
names are the names you want the columsn to have. The values are the
placeholder values used in 'git log'.}
}
\description{
Note if the benchmarks were run on a remote system you may need to fetch the
data locally first with \code{cb_fetch()}.
}
\examples{
\dontrun{
cb_read(additional_columns=c("tree_hash" = "\%T", "author_email" = "\%ae"))
}
}
\seealso{
Other cb: 
\code{\link{cb_fetch}()},
\code{\link{cb_run}()}
}
\concept{cb}
