% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\title{Benchmark a list of quoted expressions}
\usage{
mark(exprs, env = parent.frame(), n = 10, check_results = TRUE)
}
\description{
Benchmark a list of quoted expressions
}
\examples{
mark(list(quote(1 + 1), quote(2 == 2)))
}
