% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_benchmarking.R
\name{run_benchmarks}
\alias{run_benchmarks}
\alias{run_benchmark}
\title{Run benchmarks for a package}
\usage{
run_benchmarks(path = ".", env = new.env(parent = globalenv()))

run_benchmark(path, env = new.env(parent = globalenv()))
}
\arguments{
\item{path}{\code{\link[=run_benchmarks]{run_benchmarks()}} a path to a package, or within a package. For \code{\link[=run_benchmark]{run_benchmark()}} the path to the benchmark file to be run.}

\item{env}{Environment in which to execute the benchmarks.}
}
\description{
\code{\link[=run_benchmarks]{run_benchmarks()}} runs all the benchmarks in a package. \code{\link[=run_benchmark]{run_benchmark()}}
runs a single benchmark.
}
