% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{workout}
\alias{workout}
\title{Time a group of expressions individually}
\usage{
workout(expr, description = NULL)
}
\arguments{
\item{expr}{one or more expressions to workout, use \code{{}} to pass multiple
expressions.}

\item{description}{A name to call each expression}
}
\description{
Given an block of expressions in \code{{}} \code{\link[=workout]{workout()}} individually times each
expression in the group.
}
